cirrhosis of the sanity

+++ cookbooks/mu-master/recipes/init.rb
@@ -259,7 +259,6 @@ end
       find -P #{rubydir}/bin -type f -exec chmod go+rx {} \\;
     EOH
     action :nothing
-    only_if { ::Dir.exists?(gemdir) }
   end
   gem_package bundler_path do
     gem_binary gembin
@@ -276,34 +275,37 @@ end
     # XXX notify mommacat if we're *not* in chef-apply... RUNNING_STANDALONE
   end
   # Expunge old versions of knife-windows
-  Dir.glob("#{gemdir}/knife-windows-*").each { |dir|
-    next if dir.match(/\/knife-windows-(#{Regexp.quote(KNIFE_WINDOWS)})$/)
-    dir.match(/\/knife-windows-([^\/]+)$/)
-    gem_package "purge #{rubydir} knife windows #{Regexp.last_match[1]} #{gembin}" do
+  if !gemdir.nil?
+    Dir.glob("#{gemdir}/knife-windows-*").each { |dir|
+      next if dir.match(/\/knife-windows-(#{Regexp.quote(KNIFE_WINDOWS)})$/)
+      dir.match(/\/knife-windows-([^\/]+)$/)
+      gem_package "purge #{rubydir} knife windows #{Regexp.last_match[1]} #{gembin}" do
+        gem_binary gembin
+        package_name "knife-windows"
+        version Regexp.last_match[1]
+        action :remove
+        only_if { ::Dir.exists?(dir) }
+        only_if { ::Dir.exists?(gemdir) }
+      end
+      execute "rm -rf #{gemdir}/knife-windows-#{Regexp.last_match[1]}"
+    }
+
+    gem_package "#{rubydir} knife-windows #{KNIFE_WINDOWS} #{gembin}" do
       gem_binary gembin
       package_name "knife-windows"
-      version Regexp.last_match[1]
-      action :remove
-      only_if { ::Dir.exists?(dir) }
+      version KNIFE_WINDOWS
+      notifies :restart, "service[chef-server]", :delayed if rubydir == "/opt/opscode/embedded"
+      # XXX notify mommacat if we're *not* in chef-apply... RUNNING_STANDALONE
     end
-    execute "rm -rf #{gemdir}/knife-windows-#{Regexp.last_match[1]}"
-  }
 
-  gem_package "#{rubydir} knife-windows #{KNIFE_WINDOWS} #{gembin}" do
-    gem_binary gembin
-    package_name "knife-windows"
-    version KNIFE_WINDOWS
-    notifies :restart, "service[chef-server]", :delayed if rubydir == "/opt/opscode/embedded"
-    # XXX notify mommacat if we're *not* in chef-apply... RUNNING_STANDALONE
-  end
-
-  execute "Patch #{rubydir}'s knife-windows for Cygwin SSH bootstraps" do
-    cwd "#{gemdir}/knife-windows-#{KNIFE_WINDOWS}"
-    command "patch -p1 < #{MU_BASE}/lib/install/knife-windows-cygwin-#{KNIFE_WINDOWS}.patch"
-    not_if "grep -i 'locate_config_value(:cygwin)' #{gemdir}/knife-windows-#{KNIFE_WINDOWS}/lib/chef/knife/bootstrap_windows_base.rb"
-    notifies :restart, "service[chef-server]", :delayed if rubydir == "/opt/opscode/embedded"
-    only_if { ::Dir.exists?(gemdir) }
-    # XXX notify mommacat if we're *not* in chef-apply... RUNNING_STANDALONE
+    execute "Patch #{rubydir}'s knife-windows for Cygwin SSH bootstraps" do
+      cwd "#{gemdir}/knife-windows-#{KNIFE_WINDOWS}"
+      command "patch -p1 < #{MU_BASE}/lib/install/knife-windows-cygwin-#{KNIFE_WINDOWS}.patch"
+      not_if "grep -i 'locate_config_value(:cygwin)' #{gemdir}/knife-windows-#{KNIFE_WINDOWS}/lib/chef/knife/bootstrap_windows_base.rb"
+      notifies :restart, "service[chef-server]", :delayed if rubydir == "/opt/opscode/embedded"
+      only_if { ::Dir.exists?(gemdir) }
+      # XXX notify mommacat if we're *not* in chef-apply... RUNNING_STANDALONE
+    end
   end
 }
